delete image functionality added

+++ src/controllers/course.controllers.js
@@ -42,7 +42,7 @@ export const createCourse = async (req, res) => {
                 }
             );
         }
-        cloudinaryInstance.uploader.upload(req.file.path, {public_id: instructorEmail}, async (err, result) => {
+        cloudinaryInstance.uploader.upload(req.file.path, {public_id: findInstructor._id}, async (err, result) => {
             if (err) {
                 return res.status(500).json({
                     success: false,
@@ -170,6 +170,23 @@ export const updateCourse = async (req, res) => {
 export const deleteCourse = async (req, res) => {
    try {
      const id = req.params.id;
+     const course = await Course.findById({id});
+    if(!course) {
+        return res.status(400).json(
+            {
+                success: false,
+                error: "uable to fetch course details"
+            }
+        );
+    }
+    
+    cloudinaryInstance.uploader.destroy(course.instructor, function(error, result) {
+        if (error) {
+          console.error('Error deleting image:', error);
+        } else {
+          console.log('Image deleted:', result);
+        }
+      });
      const deleteId = await Course.deleteOne({ _id: id });
      if (!deleteId) {
          return res.status(500).json(
